Move javascript compilation to host
Previously, in order to keep consistent development environment
we've compiled javascript in the CI image. However we can utilise
power of pre-commmit for setting the node environment for all
contributors automatically. Instead of compiling the javascript
in the image, we can compile it via pre-commit in the host.

This can be done thanks to the new python breeze which is far more
flexible and can now add execution of compilation of the javascript when
needed and using pre-commit environments.

Thanks to that, we can vastly simplify the Dockerfiles and scripts that
are used to automatically build or signal that the assets need
recompilation. We can basically assume that the assets were prepared
outside of the image building (and breeze makes sure it happens)

The changes:

* node.js is not needed in images (neither PROD build nor CI)
* no need for multiple asset compilation scripts. All is done
  via pre-commit environment with `breeze compile-www-assets``
  command
* lint checks for UI do not need the docker image any more
  (they are also based on pre-commit environment)
* no more checks/warnings when you enter the image
* start-airflow command builds the compilation before entering
* prepare-airflow-package runs asset compilation before entering
  docker airflow building.